Revert unnecessary "[FIX] test file, remove key to avoid abusive call"
This reverts commit 6215c4e0d3983ba49c1f81fc70d67526004ad366.

It is actually intentional for the keys to be hardcoded, so that:
- example flask app can be easily run without requiring people to sign up first, it is good for demonstration purpose.
- test result can be consistent and centralized to one test account
So it is fine to leave the keys exposed. We acknowledge & are ok with the risk. It is also just a sandbox env key with no major potential harm (which anyway we can block anytime from our side if we need to).